local write to file working. need to work in react
