+++ client/src/containers/App/index.js
@@ -1,6 +1,9 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
+import { electronApp } from 'electron';
+import electron_data from 'electron-data';
+
 import {
 	BrowserRouter as Router,
 	Route
@@ -19,7 +22,71 @@ import { updateView } from '../../actions';
 import './index.css';
 
 
+let files = [
+  {
+    id: 1,
+    source: "github",
+    name: "file1",
+    createdAt: "April. 22, 2017",
+    lastModified: "April 24, 2022"
+  },
+  {
+    id: 2,
+    source: "github",
+    name: "file2",
+    createdAt: "April. 23, 2017",
+    lastModified: "April 25, 2022"
+  },
+  {
+    id: 3,
+    source: "github",
+    name: "fileHow are you",
+    createdAt: "April. 24, 2017",
+    lastModified: "April 26, 2022"
+  },
+  {
+    id: 4,
+    source: "googledrive",
+    name: "I'm great how about you.png",
+    createdAt: "April. 25, 2017",
+    lastModified: "April 27, 2022"
+  }
+];
+
 class App extends Component {
+  //electron data for local storage
+
+
+componentWillMount() {
+  
+  electron_data.config(
+    {
+      filename: 'service_data',
+      path: '/home/steven/Desktop/TestFolder',  //path to be determined later
+
+    }
+  );
+  electron_data.getOptions()
+    .then(options => {
+      console.log("what are my options? ", options);
+    });
+
+  electron_data.set('github', files)
+    .then(data => {
+      console.log('files: ', files);
+    });
+  electron_data.save()
+    .then(msg => {
+      msg = 'you have saved the file to path whatevs.';
+      console.log(msg);
+    });
+  electron_data.get('github')
+    .then(data => {
+      console.log('got the data? ', data); //retrieves 'files'
+    });
+}
+
+
 
 
   render() {
